feat(datagrid-web): move the useEffects into ProgressModal
